+++ src/abi.ts
@@ -110,6 +110,8 @@ export type AbiParameter = {
 
 export type AbiStateMutability = 'pure' | 'view' | 'nonpayable' | 'payable'
 
+export type AbiParameterKind = 'inputs' | 'outputs'
+
 export type AbiFunction = {
   /**
    * @deprecated use `pure` or `view` from {@link AbiStateMutability} instead
@@ -166,15 +168,15 @@ export type Abi = readonly (AbiFunction | AbiEvent | AbiError)[]
 export type TypedDataDomain = {
   chainId?: string | number | bigint
   name?: string
-  salt?: ResolvedConfig['BytesType']
+  salt?: ResolvedConfig['BytesType']['outputs']
   verifyingContract?: Address
   version?: string
 }
 
-// Subset of `AbiType` that excludes `tuple` and `function`
+// Subset of `AbiType` that excludes `tuple`, `function`, and dynamic aliases `int` and `uint`
 export type TypedDataType = Exclude<
   AbiType,
-  SolidityFunction | SolidityTuple | SolidityArrayWithTuple
+  SolidityFunction | SolidityTuple | SolidityArrayWithTuple | 'int' | 'uint'
 >
 
 export type TypedDataParameter = {