+++ src/examples/writeContract.test-d.ts
@@ -92,7 +92,7 @@ test('writeContract', () => {
         functionName: 'setSubnodeOwner',
         args: ['0xfoo', '0xbar', address],
       })
-      assertType<ResolvedConfig['BytesType']>(result)
+      assertType<ResolvedConfig['BytesType']['outputs']>(result)
     })
 
     test('tuple', () => {
@@ -133,7 +133,7 @@ test('writeContract', () => {
         address,
         abi: nestedTupleArrayAbi,
         functionName: 'f',
-        args: [{ a: 1, b: [2], c: [{ x: 1, y: 1 }] }, { x: 1, y: 1 }, 1n],
+        args: [{ a: 1, b: [2], c: [{ x: 1, y: 1 }] }, { x: 1n, y: 1n }, 1n],
       })
       assertType<
         readonly {