+++ src/utils.ts
@@ -1,6 +1,7 @@
 import type {
   Abi,
   AbiParameter,
+  AbiParameterKind,
   AbiStateMutability,
   AbiType,
   MBits,
@@ -27,19 +28,25 @@ import type { Merge, Tuple } from './types'
  * Does not include full array or tuple conversion. Use {@link AbiParameterToPrimitiveType} to fully convert arrays and tuples.
  *
  * @param TAbiType - {@link AbiType} to convert to TypeScript representation
+ * @param TAbiParameterKind - Optional {@link AbiParameterKind} to narrow by parameter type
  * @returns TypeScript primitive type
  */
-export type AbiTypeToPrimitiveType<TAbiType extends AbiType> =
-  PrimitiveTypeLookup<TAbiType>[TAbiType]
+export type AbiTypeToPrimitiveType<
+  TAbiType extends AbiType,
+  TAbiParameterKind extends AbiParameterKind = AbiParameterKind,
+> = PrimitiveTypeLookup<TAbiType, TAbiParameterKind>[TAbiType]
 
 // Using a map to look up types is faster, than nested conditional types
 // s/o https://twitter.com/SeaRyanC/status/1538971176357113858
-type PrimitiveTypeLookup<TAbiType extends AbiType> = {
+type PrimitiveTypeLookup<
+  TAbiType extends AbiType,
+  TAbiParameterKind extends AbiParameterKind = AbiParameterKind,
+> = {
   [_ in SolidityAddress]: ResolvedConfig['AddressType']
 } & {
   [_ in SolidityBool]: boolean
 } & {
-  [_ in SolidityBytes]: ResolvedConfig['BytesType']
+  [_ in SolidityBytes]: ResolvedConfig['BytesType'][TAbiParameterKind]
 } & {
   [_ in SolidityFunction]: `${ResolvedConfig['AddressType']}${string}`
 } & {
@@ -58,38 +65,43 @@ type PrimitiveTypeLookup<TAbiType extends AbiType> = {
 
 type GreaterThan48Bits = Exclude<MBits, 8 | 16 | 24 | 32 | 40 | 48 | ''>
 type LessThanOrEqualTo48Bits = Exclude<MBits, GreaterThan48Bits | ''>
-type DynamicBits = Exclude<MBits, GreaterThan48Bits | LessThanOrEqualTo48Bits>
+type NoBits = Exclude<MBits, GreaterThan48Bits | LessThanOrEqualTo48Bits>
 type BitsTypeLookup = {
   [_ in `${LessThanOrEqualTo48Bits}`]: ResolvedConfig['IntType']
 } & {
   [_ in `${GreaterThan48Bits}`]: ResolvedConfig['BigIntType']
 } & {
-  [_ in DynamicBits]: ResolvedConfig['IntType'] | ResolvedConfig['BigIntType']
+  [_ in NoBits]: ResolvedConfig['BigIntType']
 }
 
 /**
  * Converts {@link AbiParameter} to corresponding TypeScript primitive type.
  *
  * @param TAbiParameter - {@link AbiParameter} to convert to TypeScript representation
+ * @param TAbiParameterKind - Optional {@link AbiParameterKind} to narrow by parameter type
  * @returns TypeScript primitive type
  */
 export type AbiParameterToPrimitiveType<
   TAbiParameter extends AbiParameter | { name: string; type: unknown },
+  TAbiParameterKind extends AbiParameterKind = AbiParameterKind,
 > =
   // 1. Check to see if type is basic (not tuple or array) and can be looked up immediately.
   TAbiParameter['type'] extends Exclude<AbiType, SolidityTuple | SolidityArray>
-    ? AbiTypeToPrimitiveType<TAbiParameter['type']>
+    ? AbiTypeToPrimitiveType<TAbiParameter['type'], TAbiParameterKind>
     : // 2. Check if type is tuple and covert each component
     TAbiParameter extends {
         type: SolidityTuple
         components: infer TComponents extends readonly AbiParameter[]
       }
-    ? _HasUnnamedAbiParameter<TComponents> extends true
+    ? TComponents extends readonly []
+      ? []
+      : _HasUnnamedAbiParameter<TComponents> extends true
       ? // Has unnamed tuple parameters so return as array
         readonly [
           ...{
             [K in keyof TComponents]: AbiParameterToPrimitiveType<
-              TComponents[K]
+              TComponents[K],
+              TAbiParameterKind
             >
           },
         ]
@@ -99,7 +111,7 @@ export type AbiParameterToPrimitiveType<
             name: string
           }
             ? Component['name']
-            : never]: AbiParameterToPrimitiveType<Component>
+            : never]: AbiParameterToPrimitiveType<Component, TAbiParameterKind>
         }
     : // 3. Check if type is array.
     /**
@@ -127,11 +139,15 @@ export type AbiParameterToPrimitiveType<
         // and passed back into the function to continue reduing down to the basic types found in Step 1.
         Size extends keyof SolidityFixedArraySizeLookup
         ? Tuple<
-            AbiParameterToPrimitiveType<Merge<TAbiParameter, { type: Head }>>,
+            AbiParameterToPrimitiveType<
+              Merge<TAbiParameter, { type: Head }>,
+              TAbiParameterKind
+            >,
             SolidityFixedArraySizeLookup[Size]
           >
         : readonly AbiParameterToPrimitiveType<
-            Merge<TAbiParameter, { type: Head }>
+            Merge<TAbiParameter, { type: Head }>,
+            TAbiParameterKind
           >[]
       : never
     : // 4. If type is not basic, tuple, or array, we don't know what the type is.
@@ -153,22 +169,27 @@ type _HasUnnamedAbiParameter<TAbiParameters extends readonly AbiParameter[]> =
       ? Head['name'] extends ''
         ? true
         : _HasUnnamedAbiParameter<Tail>
-      : false
+      : true
     : false
 
 /**
  * Converts array of {@link AbiParameter} to corresponding TypeScript primitive types.
  *
  * @param TAbiParameters - Array of {@link AbiParameter} to convert to TypeScript representations
+ * @param TAbiParameterKind - Optional {@link AbiParameterKind} to narrow by parameter type
  * @returns Array of TypeScript primitive types
  */
 export type AbiParametersToPrimitiveTypes<
   TAbiParameters extends readonly AbiParameter[],
+  TAbiParameterKind extends AbiParameterKind = AbiParameterKind,
 > = {
   // TODO: Convert to labeled tuple so parameter names show up in autocomplete
   // e.g. [foo: string, bar: string]
   // https://github.com/microsoft/TypeScript/issues/44939
-  [K in keyof TAbiParameters]: AbiParameterToPrimitiveType<TAbiParameters[K]>
+  [K in keyof TAbiParameters]: AbiParameterToPrimitiveType<
+    TAbiParameters[K],
+    TAbiParameterKind
+  >
 }
 
 /**
@@ -324,10 +345,14 @@ export type IsTypedData<TTypedData> = TTypedData extends TypedData
  * Converts {@link TTypedData} to corresponding TypeScript primitive types.
  *
  * @param TTypedData - {@link TypedData} to convert
+ * @param TAbiParameterKind - Optional {@link AbiParameterKind} to narrow by parameter type
  * @returns Union of TypeScript primitive types
  */
 // TODO: Check for recursive structs (e.g. add generic slot for recursion)
-export type TypedDataToPrimitiveTypes<TTypedData extends TypedData> = {
+export type TypedDataToPrimitiveTypes<
+  TTypedData extends TypedData,
+  TAbiParameterKind extends AbiParameterKind = AbiParameterKind,
+> = {
   [K in keyof TTypedData]: {
     // Map over typed data values and turn into key-value pairs
     [K2 in TTypedData[K][number] as K2['name']]: K2['type'] extends K // 1. Eliminate self-referencing structs
@@ -347,10 +372,11 @@ export type TypedDataToPrimitiveTypes<TTypedData extends TypedData> = {
                 TTypedData
               >
             }
-          >
+          >,
+          TAbiParameterKind
         >
       : K2['type'] extends TypedDataType // 4. Known type to convert
-      ? AbiParameterToPrimitiveType<K2>
+      ? AbiParameterToPrimitiveType<K2, TAbiParameterKind>
       : `Error: Cannot convert unknown type '${K2['type']}' to primitive type.`
   }
 }